backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pymongo
motor
pydantic
pydantic-settings
python-jose[cryptography]
google-auth
python-multipart
pytest
pytest-asyncio
httpx
